feat: Create workout name for file_name
With the new `.plan` format on the wahoo element roam v3, the `name` in the file is actually the title shown on the device. Before, the file name was shown as workout name which is also why the current name is always set to "Untitled".  But now only having "Untitled" workouts is not feasable because it will be very hard to distinguish. Therefore, now I take the file name and try to parse a useful workout title from it.